🚚 Move types to shared place
